[EA3-183] refactor : 시간 투표 통계 검증 로직을 TimeVoteStatValidator 클래스로 분리 및 log 추가
